Created Factor Matrices
This are according to the description in Tammy Kolda’s Tensor Review
paper: http://www.sandia.gov/~tgkolda/pubs/pubfiles/TensorReview.pdf

These could be useful for a better csv conversion, or possibly for a
better understanding of our individual factors in visualization.

More work needs to be done on understanding the mathematics behind
these factor matrices and on ensuring the correctness of the Khatri-Rao
function in Tensorly.