Added two command line arguments: lines and axis
Lines gives an upper limit for the number of lines that the script
should read. By default this is set to 100. It would be nice to have a
“every line” option in the future.

axis: Describes the factor matrix from the decomposition that is
modeled. It is possible that the help message is not good enough for
this. Need feedback.