Remove create_factor_matrices function
Reasons
1) This function has never been used.
2) Its naming and function is confusing. What is does is attempt to
reconstruct matricizations of the original tensor from its
decomposition. It looks like it creates the factors of the tensor from
the decomposition
3) Keeping it in the code base necessitates keeping the Tensorly
dependency, or writing ur own khatri-rao function. The latter is not
too difficult, but is probably not worth the technical debt for
something we are not using.