Experimental Tensor Construction (#31)
* Removing unused import statements

Also updated gitignore

* Adding svd for matrix slices, savable tensor

Also removing imports that was supposed to happen in last commit.

* Initial movement to incorporate new tensor construction with already existing code

* Bugfix to make tensor have proper shape

* Tensor now uses only first 100 lines of file

Done for memory saving purposes

* Added file name argument

Also included temporary hack solution to visualization to support both
text and binary tensors

* Removed docSenTerm.py

This was a file for testing out the tensor before integrating it with
the rest of the code base. At this point it has served its purpose

* Added two command line arguments: lines and axis

Lines gives an upper limit for the number of lines that the script
should read. By default this is set to 100. It would be nice to have a
“every line” option in the future.

axis: Describes the factor matrix from the decomposition that is
modeled. It is possible that the help message is not good enough for
this. Need feedback.

* updated gitignore

* Using Tensorflow implementation of parafac decomposition

This does not include the new requirements, only code to use those
requirements.

* Added exception handling for large Tensors being written to pickle files

* Remove no longer needed dependencies

* Remove create_factor_matrices function

Reasons
1) This function has never been used.
2) Its naming and function is confusing. What is does is attempt to
reconstruct matricizations of the original tensor from its
decomposition. It looks like it creates the factors of the tensor from
the decomposition
3) Keeping it in the code base necessitates keeping the Tensorly
dependency, or writing ur own khatri-rao function. The latter is not
too difficult, but is probably not worth the technical debt for
something we are not using.

* Update requirements

* Remove ngrams as a required argument

* Added some explanatory comments

* Updated Readme

Adding our new team member!
Also removing outdated information